Spell check base object
